Add summarization logic

+++ src/Cargo.Data.Application/Models/RequestDto/Foo1/Crumb.cs
@@ -4,5 +4,5 @@ public class Crumb
 {
     public DateTimeOffset CreatedDtm { get; set; }
 
-    public float Value { get; set; }
+    public double Value { get; set; }
 }
